requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
base58
bech32
